requirements.txt
hermes_python>=0.1
python-crontab
playsound
